requirements.txt
streamlit
google-api-core
google-cloud
google-cloud-discoveryengine
google-cloud-storage
protobuf
google-cloud-aiplatform
google-auth
google-api-python-client
google-auth-oauthlib
finnhub-python
streamlit-google-auth
BeautifulSoup4
google-cloud-secret-manager
streamlit_float
google-genai
streamlit-pills
google-cloud-pubsub
pandas-gbq